Refactor import statements and fix type imports
